added FrontEnd with angularJS

+++ src/main/java/claudio/happyhour/model/Worker.java
@@ -22,7 +22,7 @@ public class Worker {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private long id;
 
     private String email;
     private String name;
@@ -41,11 +41,11 @@ public class Worker {
         }
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
